increase Value range to +/-15999

+++ util/src/value.rs
@@ -6,7 +6,7 @@ pub struct ValueBounds;
 impl Bounds for ValueBounds {
     type Integer = i16;
     const LOWER: Self::Integer = -Self::UPPER;
-    const UPPER: Self::Integer = ScoreBounds::UPPER - PlyBounds::UPPER as i16 - 1;
+    const UPPER: Self::Integer = ScoreBounds::UPPER - PlyBounds::UPPER - 1;
 }
 
 /// A position's static evaluation.
